tests working! basic abstraction done, better abstraction needed
The api to using a typeahead only takes a function with which to parse
the input text and some placeholder text.  An improvement might be to
take a map that is merged into the properties of the input hiccup
item.

Then we could pass in things like styling information etc...